Editor classes + Cache revised * added developer-entities to the project-cache * added count of solutions and relatiosn to project-versions * from now on you can only delete versions without solutions and relations * added method for getting developer-entities to Project * revised the complete delete-mechanism now it should reliable delete everything

+++ files/lib/acp/action/VersionDeleteAction.class.php
@@ -19,7 +19,10 @@ class VersionDeleteAction extends AbstractVersionAction {
 	 * Deletes the version.
 	 */
 	public function action() {
-		$this->version->delete();
+		if ($this->version->solutions == 0 && $this->version->relations == 0) {
+			$this->version->delete();
+		}
+		else throw new IllegalLinkException();
 	}
 }
 ?>
